--- 61030060/CL62-03/Hello/Hello/Program.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Hello
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello Computer Lad");
-            Console.ReadKey();
-        }
-    }
-}
